solve cycle been dependency problem, fix some bugs

+++ src/main/java/com/certTrack/UserService/Repository/UserRepository.java
@@ -1,10 +1,13 @@
 package com.certTrack.UserService.Repository;
 
 import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import com.certTrack.UserService.Entity.User;
 
-public interface UserRepository extends JpaRepository<User, Integer>{
+public interface UserRepository extends JpaRepository<User, Long>{
 
 	Optional<User> findByEmail(String email);
+
 }
